+++ buildSrc/src/main/groovy/org/apache/beam/gradle/BeamModulePlugin.groovy
@@ -321,19 +321,17 @@ class BeamModulePlugin implements Plugin<Project> {
 
   // A class defining the common properties in a given suite of cross-language tests
   // Properties are shared across runners and are used when creating a CrossLanguageUsingJavaExpansionConfiguration object
-  static class CrossLanguageTaskCommon {
+  static class CrossLanguageTask {
     // Used as the task name for cross-language
     String name
-    // The expansion service's project path (required)
-    String expansionProjectPath
+    // List of project paths for required expansion services
+    List<String> expansionProjectPaths
     // Collect Python pipeline tests with this marker
     String collectMarker
-    // Job server startup task.
-    TaskProvider startJobServer
-    // Job server cleanup task.
-    TaskProvider cleanupJobServer
-    // any additional environment variables specific to the suite of tests
+    // Additional environment variables to set before running tests
     Map<String,String> additionalEnvs
+    // Additional Python dependencies to install before running tests
+    List<String> additionalDeps
   }
 
   // A class defining the configuration for CrossLanguageUsingJavaExpansion.
@@ -349,18 +347,16 @@ class BeamModulePlugin implements Plugin<Project> {
     ]
     // Additional pytest options
     List<String> pytestOptions = []
-    // Job server startup task.
-    TaskProvider startJobServer
-    // Job server cleanup task.
-    TaskProvider cleanupJobServer
     // Number of parallel test runs.
     Integer numParallelTests = 1
-    // Project path for the expansion service to start up
-    String expansionProjectPath
+    // List of project paths for required expansion services
+    List<String> expansionProjectPaths
     // Collect Python pipeline tests with this marker
     String collectMarker
     // any additional environment variables to be exported
     Map<String,String> additionalEnvs
+    // Additional Python dependencies to install before running tests
+    List<String> additionalDeps
   }
 
   // A class defining the configuration for CrossLanguageValidatesRunner.
@@ -2576,7 +2572,7 @@ class BeamModulePlugin implements Plugin<Project> {
     /** ***********************************************************************************************/
     // Method to create the createCrossLanguageUsingJavaExpansionTask.
     // The method takes CrossLanguageUsingJavaExpansionConfiguration as parameter.
-    // This method creates a task that runs Python SDK pipeline tests that use Java transforms via an input expansion service
+    // This method creates a task that runs Python SDK test-suites that use external Java transforms
     project.ext.createCrossLanguageUsingJavaExpansionTask = {
       // This task won't work if the python build file doesn't exist.
       if (!project.project(":sdks:python").buildFile.exists()) {
@@ -2586,49 +2582,29 @@ class BeamModulePlugin implements Plugin<Project> {
       def config = it ? it as CrossLanguageUsingJavaExpansionConfiguration : new CrossLanguageUsingJavaExpansionConfiguration()
 
       project.evaluationDependsOn(":sdks:python")
-      project.evaluationDependsOn(config.expansionProjectPath)
+      for (path in config.expansionProjectPaths) {
+        project.evaluationDependsOn(path)
+      }
       project.evaluationDependsOn(":sdks:java:extensions:python")
 
-      // Setting up args to launch the expansion service
       def pythonDir = project.project(":sdks:python").projectDir
-      def javaExpansionPort = -1 // will be populated in setupTask
-      def expansionJar = project.project(config.expansionProjectPath).shadowJar.archivePath
-      def javaClassLookupAllowlistFile = project.project(config.expansionProjectPath).projectDir.getPath()
-      def expansionServiceOpts = [
-        "group_id": project.name,
-        "java_expansion_service_jar": expansionJar,
-        "java_expansion_service_allowlist_file": javaClassLookupAllowlistFile,
-      ]
       def usesDataflowRunner = config.pythonPipelineOptions.contains("--runner=TestDataflowRunner") || config.pythonPipelineOptions.contains("--runner=DataflowRunner")
       def javaContainerSuffix = getSupportedJavaVersion()
 
-      // 1. Builds the chosen expansion service jar and launches it
-      def setupTask = project.tasks.register(config.name+"Setup") {
-        dependsOn ':sdks:java:container:' + javaContainerSuffix + ':docker'
-        dependsOn project.project(config.expansionProjectPath).shadowJar.getPath()
-        dependsOn 'installGcpTest'
+      // Sets up, collects, and runs Python pipeline tests
+      project.tasks.register(config.name+"PythonUsingJava") {
+        group = "Verification"
+        description = "Runs Python SDK pipeline tests that use a Java expansion service"
+        // Each expansion service we use needs to be built before running these tests
+        // The built jars will be started up automatically using the BeamJarExpansionService utility
+        for (path in config.expansionProjectPaths) {
+          dependsOn project.project(path).shadowJar.getPath()
+        }
+        dependsOn ":sdks:java:container:$javaContainerSuffix:docker"
+        dependsOn "installGcpTest"
         if (usesDataflowRunner) {
           dependsOn ":sdks:python:test-suites:dataflow:py${project.ext.pythonVersion.replace('.', '')}:initializeForDataflowJob"
         }
-        doLast {
-          project.exec {
-            // Prepare a port to use for the expansion service
-            javaExpansionPort = getRandomPort()
-            expansionServiceOpts.put("java_port", javaExpansionPort)
-            // setup test env
-            def serviceArgs = project.project(':sdks:python').mapToArgString(expansionServiceOpts)
-            executable 'sh'
-            args '-c', ". ${project.ext.envdir}/bin/activate && $pythonDir/scripts/run_expansion_services.sh stop --group_id ${project.name} && $pythonDir/scripts/run_expansion_services.sh start $serviceArgs"
-          }
-        }
-      }
-
-      // 2. Sets up, collects, and runs Python pipeline tests
-      def pythonTask = project.tasks.register(config.name+"PythonUsingJava") {
-        group = "Verification"
-        description = "Runs Python SDK pipeline tests that use a Java expansion service"
-        dependsOn setupTask
-        dependsOn config.startJobServer
         doLast {
           def beamPythonTestPipelineOptions = [
             "pipeline_opts": config.pythonPipelineOptions + (usesDataflowRunner ? [
@@ -2641,29 +2617,19 @@ class BeamModulePlugin implements Plugin<Project> {
           def cmdArgs = project.project(':sdks:python').mapToArgString(beamPythonTestPipelineOptions)
 
           project.exec {
-            environment "EXPANSION_JAR", expansionJar
-            environment "EXPANSION_PORT", javaExpansionPort
-            for (envs in config.additionalEnvs){
-              environment envs.getKey(), envs.getValue()
+            // environment variable to indicate that jars have been built
+            environment "EXPANSION_JARS", config.expansionProjectPaths
+            String additionalDependencyCmd = ""
+            if (config.additionalDeps != null && !config.additionalDeps.isEmpty()){
+              additionalDependencyCmd = "&& pip install ${config.additionalDeps.join(' ')} "
             }
             executable 'sh'
-            args '-c', ". ${project.ext.envdir}/bin/activate && cd $pythonDir && ./scripts/run_integration_test.sh $cmdArgs"
+            args '-c', ". ${project.ext.envdir}/bin/activate " +
+                additionalDependencyCmd +
+                "&& cd $pythonDir && ./scripts/run_integration_test.sh $cmdArgs"
           }
         }
       }
-
-      // 3. Shuts down the expansion service
-      def cleanupTask = project.tasks.register(config.name+'Cleanup', Exec) {
-        // teardown test env
-        executable 'sh'
-        args '-c', ". ${project.ext.envdir}/bin/activate && $pythonDir/scripts/run_expansion_services.sh stop --group_id ${project.name}"
-      }
-
-      setupTask.configure {finalizedBy cleanupTask}
-      config.startJobServer.configure {finalizedBy config.cleanupJobServer}
-
-      cleanupTask.configure{mustRunAfter pythonTask}
-      config.cleanupJobServer.configure{mustRunAfter pythonTask}
     }
 
     /** ***********************************************************************************************/